refactor: adding type length

+++ src/libraries/AppStorage.sol
@@ -23,7 +23,7 @@ struct UserData {
     uint256 lastBoostClaimTimestamp;
     uint256 unlockAmount;
     uint256 unlockTimestamp;
-    uint amountClaimed;
+    uint256 amountClaimed;
     bool hasWithdrawnOrRestaked;
     
 }
